pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sustainability-gleif-matching-network"
version = "0.2.0"
description = "GLEIF LEI/RR parsing, sustainability-source assembly, fuzzy matching, and ownership network outputs"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.0",
  "rapidfuzz>=3.6.0",
  "pyarrow>=15.0.0",
  "PyYAML>=6.0.0",
  "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
